requirements.txt
python-dateutil>=2.9.0
pydantic>=2.8.2
requests>=2.32.3
networkx>=3.4.2
rapidfuzz>=3.10.1
rank-bm25>=0.2.2
streamlit>=1.42.0
pandas>=2.2.3
plotly>=5.24.1
pytest>=9.0.0
